+++ modules/data_analysis/builtin/cpp/allbetweenBuiltin.cpp
@@ -1,0 +1,26 @@
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// SPDX-License-Identifier: LGPL-3.0-or-later
+// LICENCE_BLOCK_END
+//=============================================================================
+#include "allbetweenBuiltin.hpp"
+#include "IsBetween.hpp"
+#include "InputOutputArgumentsCheckers.hpp"
+//=============================================================================
+using namespace Nelson;
+//=============================================================================
+ArrayOfVector
+Nelson::DataAnalysisGateway::allbetweenBuiltin(
+    Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
+{
+    nargincheck(argIn, 3);
+    nargoutcheck(nLhs, 0, 1);
+    ArrayOfVector retval;
+    retval << AllBetween(eval, argIn);
+    return retval;
+}
+//=============================================================================