+++ modules/validators/src/include/Validators.hpp
@@ -111,6 +111,8 @@ NLSVALIDATORS_IMPEXP void
 mustBeInRange(const ArrayOfVector& args, const ArrayOf& upper, const ArrayOf& lower,
     const std::wstring& boundflag1, const std::wstring& boundflag2, int argPosition);
 //=============================================================================
-
+NLSVALIDATORS_IMPEXP void
+mustBeBetween(const ArrayOfVector& args, int argPosition);
+//=============================================================================
 }
 //=============================================================================